Fix calendar and calendar_event duplication bugs
Fixes #850
Fixes #857
